File structure reorg to Arduino Library Reference
