+++ library.properties
@@ -1,0 +1,10 @@
+name=ArTICL
+version=0.0.1
+author=KermMartian
+maintainer=KermMartian
+sentance=Arduino TI Calculator Linking Library (ArTICL) for interfacing with TI-8x/9x calculators over the TI Link Protocol
+paragraph=Support for the TI graphing calculators with the 2.5mm I/O port.
+category=Communication
+url=https://github.com/KermMartian/ArTICL
+architectures=*
+includes=ArTICL.h
