+++ src/ArTICL.h
@@ -1,0 +1,8 @@
+#ifndef ARTCIL_H
+#define ARTCIL_H
+
+#include "CBL2.h"
+#include "TICL.h"
+#include "TIVar.h"
+
+#endif	// ARTICL_H
